rename advisor cockpit snapshot router
