+++ src/app/routers/advisor_cockpit_snapshot.py
